move SecretHeader.decorator to decorator folder
